refactor: Organize code into subfolders
